Airspeed: allow adjustment of differential pressure scaling
The current code assumes that the sensor used is an MS4525D0 1PSI
model. There are many differently-scaled MS4525 sensors, and also
a variety of Honeywell ones (HSC/SSC series) that have the same
interface (ie I2C address 0x28) but different scaling. Since all
the sensors have the same address it's not possible to
automatically determine which one is connected.

The code has been changed so that the scaling can be adjusted
manually with a call to ioctl. On startup it defaults to the
scaling used for the 1 PSI sensor.